Extract currency REST exception handler from domain advice
